Split off the vision-specific code from wrappers.py
