+++ conduit/data/datasets/vision/__init__.py
@@ -11,3 +11,4 @@ from .pacs import *
 from .ssrp import *
 from .utils import *
 from .waterbirds import *
+from .wrappers import *
